@@ -30,7 +30,7 @@
         ]
     },
     "require": {
-        "php": ">=5.6.0",
+        "php": ">=7.2.0",
         "guzzlehttp/guzzle":"~6",
         "webignition/url":"~2",
         "webignition/web-resource-retriever":">=0.7,<1",
@@ -38,8 +38,8 @@
         "webignition/robots-txt-file":">=2.0,<3"
     },
     "require-dev": {
-        "phpunit/phpunit": "~5.0",
-        "squizlabs/php_codesniffer": "3.*"
+        "phpunit/phpunit": "^7",
+        "squizlabs/php_codesniffer": "^3"
     },
     "minimum-stability":"stable",
     "prefer-stable":true
